requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
flatbencode==0.2.1
idna==3.4
python-dotenv==1.0.0
qbittorrent-api==2023.4.47
requests==2.28.2
six==1.16.0
torf==4.2.0
urllib3==1.26.15
